Normalized config to use @brisberg/typescript-toolchain (#11)
Merged in @brisberg/typescript-toolchain. TS-Toolchain v2.0.2

This normalized the configuration of TypeScript ecosystem tools to the standard for @brisberg projects.

Also upgraded @types/screeps to the latest pre-release version to take advantage of Decorations Update typings.

Also pinned screeps-jest to my fork on the branch which fixed compatibility with Jest@26.